🔥fix: 오늘의 섭취량 데이터 타입 수정

+++ android-app/app/src/main/java/com/swallaby/foodon/presentation/main/component/NutrientIntakeContent.kt
@@ -65,9 +65,9 @@ fun NutrientIntakeContent(
                 val fatRatio = it.intakeFat.toFloat() * 9 / it.goalKcal
 
                 listOf(
-                    Nutrition(NutritionType.CARBOHYDRATE, it.intakeCarbs.toDouble(), carbsRatio),
-                    Nutrition(NutritionType.PROTEIN, it.intakeProtein.toDouble(), proteinRatio),
-                    Nutrition(NutritionType.FAT, it.intakeFat.toDouble(), fatRatio),
+                    Nutrition(NutritionType.CARBOHYDRATE, it.intakeCarbs, carbsRatio),
+                    Nutrition(NutritionType.PROTEIN, it.intakeProtein, proteinRatio),
+                    Nutrition(NutritionType.FAT, it.intakeFat, fatRatio),
                 ).sortedByDescending { n -> n.amount }
             } ?: emptyList()
